:sparkles: Paginated activity feed
